projects/std.flow.minimal/requirements.txt
signac>=0.8
signac-flow>=0.5
